move the authz code around
