+++ plugins/nemo-automodel/src/nemo_automodel_plugin/transform.py
@@ -8,8 +8,8 @@ from __future__ import annotations
 import uuid
 from typing import TYPE_CHECKING
 
+from nemo_platform_plugin.refs import parse_entity_ref
 from nmp.automodel.platform_client import check_dataset_access, fetch_model_entity
-from nmp.common.entities.utils import parse_entity_ref
 
 from nemo_automodel_plugin.schema import (
     AutomodelJobInput,
