+++ tools/lint/lint-no-nmp-common-in-plugins.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+set -euo pipefail
+uv run pre-commit run no-nmp-common-in-plugins --all-files
